Improve handling of WiFi connection errors
Sometimes the unit cannot connect to WiFi and in this case it just keeps trying,
quickly draining the battery. So introduce a counter for the number of
connection attempts, and go to sleep if this counter reaches a predefined
number of attempts.
This counter is also exported as Prometheus metric at conn_failures_count.